feat: add another menu for the saved configs
